requirements.txt
dnspython>=2.2.0
geoip2>=4.2.0
cachetools>=5.3.0
requests>=2.27.1
aiohttp>=3.8.0
pydantic>=1.9.0
click>=8.0.0
rich>=12.0.0
python-dotenv>=0.20.0 
openai>=1.1.0
cryptography